Fix missing cast

+++ docs/src/lib.rs
@@ -305,7 +305,7 @@ pub fn files_to_documentations(
             &std_lib,
             src_dir.as_path(),
             MutMap::default(),
-            std::mem::size_of::<usize>(), // This is just type-checking for docs, so "target" doesn't matter
+            std::mem::size_of::<usize>() as u32, // This is just type-checking for docs, so "target" doesn't matter
             builtin_defs_map,
         ) {
             Ok(loaded) => files_docs.push((loaded.documentation, loaded.interns)),
